# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SpaceArt API",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://api-spaceart.local/api/"
        }
    ],
    "paths": {
...
        "/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "List random posts",
                "description": "Fetch random posts on database",
                "operationId": "425e95aaa6ab243fc407ffcf09db9de5",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts were found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Post"
                                            }
                                        },
                                        "fails": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Publish post",
                "description": "Store post",
                "operationId": "35cc3a4123556e20dc17f6f45be00f18",
                "requestBody": {
                    "$ref": "#/components/requestBodies/PostStore"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/ReturnPost"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/post/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Fetch unique post",
                "description": "Fetch post on database",
                "operationId": "820a95132e64f02c9c71b2cfc3cadb0b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReturnPost"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete post",
                "description": "Delete post on database",
                "operationId": "cdf3a5c9a54637677e9d19e0bd933ab4",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "Post was deleted"
                                        },
                                        "fails": {
                                            "type": "boolean",
                                            "default": "false"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
...
        "/agreement/{agreement}/rate": {
            "post": {
                "tags": [
                    "Rate"
                ],
                "summary": "Store rate",
                "description": "Relates rate to agreement on database",
                "operationId": "b31d47e125d7ced6cba151fdeaee3a31",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Agreement"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/RateStore"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/ReturnRate"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/agreement/{agreement}/rate/{author}": {
            "get": {
                "tags": [
                    "Rate"
                ],
                "summary": "Fetch rate",
                "description": "Fetch unique rate of an agreement",
                "operationId": "f7782f6db85bd524b92a5f6e9165a240",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Agreement"
                    },
                    {
                        "$ref": "#/components/parameters/Author"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReturnRate"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rate"
                ],
                "summary": "Update rate",
                "description": "Update rate on database",
                "operationId": "264c601c7f38b30f4c67b7eb0cad92d8",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Agreement"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/RateUpdate"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReturnRate"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rate"
                ],
                "summary": "Delete rate",
                "description": "Delete rate on database",
                "operationId": "6a7bc5251550e59c502647c1e8e0e4d6",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Agreement"
                    },
                    {
                        "$ref": "#/components/parameters/Author"
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/agreement/{agreement}/rate/{author}/update": {
            "post": {
                "tags": [
                    "Rate"
                ],
                "summary": "[PUT]::/agreement/{agreement}/rate/{author} alias",
                "description": "Redirect request to [PUT]::/agreement/{agreement}/rate/{author}",
                "operationId": "11277d33f03ad19571937f528f5ce8b0",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Agreement"
                    },
                    {
                        "$ref": "#/components/parameters/Author"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirected to [PUT]::/agreement/{agreement}/rate/{author}"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/agreement/{agreement}/rate/{author}/delete": {
            "post": {
                "tags": [
                    "Rate"
                ],
                "summary": "[DELETE]::/agreement/{agreement}/rate/{author} alias",
                "description": "Redirect request to [DELETE]::/agreement/{agreement}/rate/{author}",
                "operationId": "30aef47e3a00200df4f8abd616c3b96d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Agreement"
                    },
                    {
                        "$ref": "#/components/parameters/Author"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirected to [DELETE]::/agreement/{agreement}/rate/{author}"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Post": {
                "description": "Schema of Model Post",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "text": {
                        "description": "Text",
                        "type": "string",
                        "example": "Today i show me in Atlantic Resort"
                    },
                    "image": {
                        "description": "Image in base64 or url",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Rate": {
                "description": "Schema of Model Rate",
                "properties": {
                    "author": {
                        "$ref": "#/components/schemas/User"
                    },
                    "rated": {
                        "$ref": "#/components/schemas/User"
                    },
                    "agreement": {
                        "$ref": "#/components/schemas/Agreement"
                    },
                    "score": {
                        "type": "number",
                        "format": "float",
                        "maximum": 5,
                        "minimum": 0,
                        "example": "4.5"
                    },
                    "note": {
                        "description": "Short review",
                        "type": "string",
                        "example": "The artist is very helpful"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "description": "Schema of Model User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Username",
                        "type": "string",
                        "example": "José Carlos"
                    },
                    "active": {
                        "description": "Access to account is enabled",
                        "type": "boolean",
                        "default": "true"
                    },
                    "email": {
                        "description": "Access email address",
                        "type": "string",
                        "example": "example@org.net"
                    },
                    "phone": {
                        "description": "User's mobile number",
                        "type": "string",
                        "example": "71988469787"
                    },
                    "type": {
                        "description": "Account type",
                        "type": "enum",
                        "enum": [
                            "artist",
                            "enterprise"
                        ]
                    },
                    "postal_code": {
                        "description": "Brazilian zip code",
                        "type": "string",
                        "example": "41000000"
                    },
                    "state": {
                        "description": "Brazilian state acronym",
                        "type": "string",
                        "example": "BA"
                    },
                    "city": {
                        "type": "string",
                        "example": "Salvador"
                    },
                    "neighborhood": {
                        "type": "string",
                        "example": "Piatã"
                    },
                    "street": {
                        "type": "string",
                        "example": "Av. Orlando Gomes"
                    },
                    "address_complement": {
                        "description": "Additional  information of address",
                        "type": "string",
                        "example": "Beside of SESI Saúde"
                    },
                    "image": {
                        "description": "Image in base64 or URL",
                        "type": "string"
                    },
                    "slug": {
                        "description": "User's website address",
                        "type": "string",
                        "example": "https://spaceart-lemon.vercel.app/user/2",
                        "nullable": true
                    },
                    "biography": {
                        "description": "Short presentation text",
                        "type": "string",
                        "nullable": true
                    },
                    "received_rates_avg_score": {
                        "type": "number",
                        "nullable": true
                    },
                    "artist_account_data": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Artist"
                            }
                        ],
                        "nullable": true
                    },
                    "enterprise_account_data": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Enterprise"
                            }
                        ],
                        "nullable": true
                    },
                    "sent_rates": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rate"
                        }
                    },
                    "received_rates": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Rate"
                        }
                    },
                    "posts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Post"
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "204": {
                "description": "Resource was disabled",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "fails": {
                                    "type": "boolean",
                                    "default": "false"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ReturnPost": {
                "description": "Operation finished succesfully",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Post was created"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                },
                                "fails": {
                                    "type": "boolean",
                                    "default": "false"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ReturnRate": {
                "description": "Response is successful",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rate"
                                    },
                                    "maxItems": 1
                                },
                                "fails": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "Limit": {
                "name": "limit",
                "in": "query",
                "description": "Limit per page",
                "style": "form",
                "schema": {
                    "type": "integer",
                    "nullable": true
                }
            },
...
            "Id": {
                "name": "id",
                "in": "path",
                "description": "Resource id",
                "style": "form",
                "schema": {
                    "type": "integer"
                }
            },
            "Author": {
                "name": "author",
                "in": "path",
                "description": "Author id",
                "style": "form",
                "schema": {
                    "type": "integer"
                }
            },
            "Agreement": {
                "name": "agreement",
                "in": "path",
                "description": "Agreement id",
                "style": "form",
                "schema": {
                    "type": "integer"
                }
            }
        },
        "requestBodies": {
...
            "PostStore": {
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "user_id": {
                                    "type": "integer"
                                },
                                "text": {
                                    "type": "string"
                                },
                                "image": {
                                    "description": "Image in base64 ou URL",
                                    "type": "string",
                                    "maxLength": 10000000
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RateStore": {
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "author_id",
                                "score",
                                "note"
                            ],
                            "properties": {
                                "author_id": {
                                    "type": "integer"
                                },
                                "score": {
                                    "type": "number",
                                    "maximum": 5,
                                    "minimum": 0
                                },
                                "note": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "RateUpdate": {
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "score"
                            ],
                            "properties": {
                                "score": {
                                    "type": "number",
                                    "maximum": 5,
                                    "minimum": 0
                                },
                                "note": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Sanctum": {
                "type": "apiKey",
                "scheme": {
                    "bearerToken": []
                }
            }
        }
    }
}